requirements.txt
APScheduler==3.10.1
black==23.11.0
click==8.1.3
datasets==2.14.5
gradio==4.4.0
gradio_client==0.7.0
huggingface-hub>=0.18.0
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
python-dateutil==2.8.2
requests==2.28.2
tqdm==4.65.0
transformers==4.35.2
tokenizers>=0.15.0
git+https://github.com/EleutherAI/lm-evaluation-harness.git@b281b0921b636bc36ad05c0b0b0763bd6dd43463#egg=lm-eval
accelerate==0.24.1